+++ src/Services/Profiles/Profiles.Services/Services/ReceptionistsService.cs
@@ -1,12 +1,12 @@
 ﻿using AutoMapper;
 using Microsoft.IdentityModel.Tokens;
+using OneOf;
+using OneOf.Types;
 using Profiles.Contracts.DTOs.OuterServicesModels;
 using Profiles.Contracts.DTOs.Receptionist;
 using Profiles.Domain.Entities;
-using Profiles.Domain.Exceptions;
 using Profiles.Domain.Interfaces;
 using Profiles.Services.Abstractions;
-using System.Net.Http.Json;
 
 namespace Profiles.Services.Services;
 
@@ -14,69 +14,52 @@ public class ReceptionistsService : IReceptionistsService
 {
     private readonly IRepositoryManager _repositoryManager;
     private readonly IMapper _mapper;
-    private readonly HttpClient _httpClient;
+    private readonly IHttpRepository<OfficeDTO> _httpRepository;
 
-    public ReceptionistsService(IRepositoryManager repositoryManager, IMapper mapper, IHttpClientFactory factory)
+    public ReceptionistsService(IRepositoryManager repositoryManager, 
+        IMapper mapper, IHttpRepository<OfficeDTO> httpRepository)
     {
         _repositoryManager = repositoryManager;
         _mapper = mapper;
-        _httpClient = factory.CreateClient();
-        _httpClient.BaseAddress = new Uri("https://localhost:7255/api/offices/");
+        _httpRepository = httpRepository;
     }
 
-    public async Task<List<ReceptionistResponseDTO>> GetAllReceptionistsAsync(bool trackChanges)
+    public async Task<OneOf<List<ReceptionistResponseDTO>, NotFound>> GetAllReceptionistsAsync(bool trackChanges)
     {
         var receptionists = await _repositoryManager.ReceptionistsRepository.GetAllAsync(trackChanges);
 
         if (receptionists.IsNullOrEmpty())
         {
-            throw new NotFoundException("There are not receptionists in the database.");
+            return new NotFound();
         }
 
         var mappedReceptionists = _mapper.Map<List<ReceptionistResponseDTO>>(receptionists);
 
-        try
+        IEnumerable<string> officesIds = receptionists.Select(x => x.OfficeId).Distinct().ToList();
+
+        var offices = await _httpRepository.GetCollection(officesIds);
+
+        foreach (var receptionist in mappedReceptionists)
         {
-            IEnumerable<string> officesIds = receptionists.Select(x => x.OfficeId).Distinct().ToList();
-
-            string stringWithOfficesIds = string.Join(',', officesIds);
-
-            var offices = await _httpClient.GetFromJsonAsync<List<OfficeDTO>>($"collection/({stringWithOfficesIds})");
-
-            foreach (var receptionist in mappedReceptionists)
-            {
-                receptionist.Office = offices.First(x => x.OfficeId.Equals(receptionist.Office.OfficeId));
-            }
-        }
-        catch (Exception) 
-        {
-            throw new Exception("Something went wrong during the request to OfficeAPI");
+            receptionist.Office = offices.First(x => x.OfficeId.Equals(receptionist.Office.OfficeId));
         }
 
         return mappedReceptionists;
     }
 
-    public async Task<ReceptionistResponseDTO?> GetReceptionistByIdAsync(Guid id, bool trackChanges)
+    public async Task<OneOf<ReceptionistResponseDTO, NotFound>> GetReceptionistByIdAsync(Guid id, bool trackChanges)
     {
         var receptionist = await _repositoryManager.ReceptionistsRepository.GetByIdAsync(id, trackChanges);
 
         if (receptionist is null)
         {
-            throw new NotFoundException($"Receptionist with id: {id} was not found in the database.");
+            return new NotFound();
         }
 
         var mappedReceptionist = _mapper.Map<ReceptionistResponseDTO>(receptionist);
 
-        try
-        {
-            var office = await _httpClient.GetFromJsonAsync<OfficeDTO>($"{mappedReceptionist.Office.OfficeId}");
-
-            mappedReceptionist.Office = office;
-        }
-        catch (Exception)
-        {
-            throw new Exception("Something went wrong during the request to OfficeAPI");
-        }
+            mappedReceptionist.Office = 
+                await _httpRepository.GetOneAsync("https://localhost:7255/api/offices", receptionist.OfficeId!);
 
         return mappedReceptionist;
     }
@@ -106,28 +89,20 @@ public class ReceptionistsService : IReceptionistsService
         await _repositoryManager.SaveAsync();
         
         var receptionistResult = _mapper.Map<ReceptionistResponseDTO>(newReceptionistEntity);
-        
-        try
-        {
-            var office = await _httpClient.GetFromJsonAsync<OfficeDTO>($"{newReceptionist.OfficeId}");
 
-            receptionistResult.Office = office;
-        }
-        catch (Exception)
-        {
-            throw new Exception("Something went wrong during the request to OfficeAPI");
-        }
+        receptionistResult.Office =  
+            await _httpRepository.GetOneAsync("https://localhost:7255/api/offices", newReceptionistEntity.OfficeId!);
 
         return receptionistResult;
     }
 
-    public async Task UpdateReceptionistAsync(Guid id, ReceptionistUpdateDTO updatedReceptionist)
+    public async Task<OneOf<Success, NotFound>> UpdateReceptionistAsync(Guid id, ReceptionistUpdateDTO updatedReceptionist)
     {
         var receptionistEntity = await _repositoryManager.ReceptionistsRepository.GetByIdAsync(id, true);
 
         if (receptionistEntity is null)
         {
-            throw new NotFoundException($"The receptionist with id: {id} was not found in the database.");
+            return new NotFound();
         }
 
         _mapper.Map(updatedReceptionist, receptionistEntity);
@@ -135,14 +110,23 @@ public class ReceptionistsService : IReceptionistsService
         receptionistEntity.OfficeId = updatedReceptionist.OfficeId;
 
         await _repositoryManager.SaveAsync();
+
+        return new Success();
     }
 
-    public async Task DeleteReceptionistAsync(Guid id)
+    public async Task<OneOf<Success, NotFound>> DeleteReceptionistAsync(Guid id)
     {
         var receptionistEntity = await _repositoryManager.ReceptionistsRepository.GetByIdAsync(id, true);
+
+        if (receptionistEntity is null)
+        {
+            return new NotFound();
+        }
 
         _repositoryManager.AccountsRepository.Delete(receptionistEntity.Account);
 
         await _repositoryManager.SaveAsync();
+
+        return new Success();
     }
 }
